feat: UI: add user page: useIframeShowMail && mailboxSplitSize
