+++ frontend/src/router/index.js
@@ -1,9 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Index from '../views/Index.vue'
-import Settings from '../views/Settings.vue'
+import User from '../views/User.vue'
 import SendMail from '../views/send/SendMail.vue'
 import Admin from '../views/Admin.vue'
-import SendBox from '../views/send/SendBox.vue'
 
 const router = createRouter({
     history: createWebHistory(),
@@ -13,16 +12,12 @@ const router = createRouter({
             component: Index
         },
         {
-            path: '/settings',
-            component: Settings
+            path: '/user',
+            component: User
         },
         {
             path: '/send',
             component: SendMail
-        },
-        {
-            path: '/sendbox',
-            component: SendBox
         },
         {
             path: '/admin',
